+++ docshell/base/BrowsingContext.cpp
@@ -131,12 +131,10 @@ already_AddRefed<BrowsingContext> BrowsingContext::Create(
   context->mName = aName;
   context->mOpenerId = aOpener ? aOpener->Id() : 0;
 
-  if (aParent) {
-    context->mCrossOriginPolicy = aParent->mCrossOriginPolicy;
-  } else if (aOpener) {
-    context->mCrossOriginPolicy = aOpener->mCrossOriginPolicy;
-  } else {
-    context->mCrossOriginPolicy = nsILoadInfo::CROSS_ORIGIN_POLICY_NULL;
+  BrowsingContext* inherit = aParent ? aParent : aOpener;
+  if (inherit) {
+    context->mOpenerPolicy = inherit->mOpenerPolicy;
+    context->mCrossOriginPolicy = inherit->mCrossOriginPolicy;
   }
 
   Register(context);