+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/CrashReporter.java
@@ -72,9 +72,9 @@ public class CrashReporter {
      * @see GeckoRuntime#ACTION_CRASHED
      */
     @AnyThread
-    public static GeckoResult<String> sendCrashReport(@NonNull final Context context,
-                                                      @NonNull final Intent intent,
-                                                      @NonNull final String appName)
+    public static @NonNull GeckoResult<String> sendCrashReport(@NonNull final Context context,
+                                                               @NonNull final Intent intent,
+                                                               @NonNull final String appName)
             throws IOException, URISyntaxException {
         return sendCrashReport(context, intent.getExtras(), appName);
     }