+++ netwerk/protocol/http/ClassifierDummyChannel.cpp
@@ -551,12 +551,6 @@ ClassifierDummyChannel::SetTopWindowURIIfUnknown(nsIURI* aTopWindowURI) {
 }
 
 NS_IMETHODIMP
-ClassifierDummyChannel::SetTopWindowPrincipal(
-    nsIPrincipal* aTopWindowPrincipal) {
-  return NS_ERROR_NOT_IMPLEMENTED;
-}
-
-NS_IMETHODIMP
 ClassifierDummyChannel::GetProxyURI(nsIURI** aProxyURI) {
   return NS_ERROR_NOT_IMPLEMENTED;
 }
@@ -623,6 +617,11 @@ ClassifierDummyChannel::CancelByChannelClassifier(nsresult aErrorCode) {
 void ClassifierDummyChannel::SetIPv4Disabled() {}
 
 void ClassifierDummyChannel::SetIPv6Disabled() {}
+
+NS_IMETHODIMP ClassifierDummyChannel::GetCrossOriginOpenerPolicy(
+    nsILoadInfo::CrossOriginOpenerPolicy* aPolicy) {
+  return NS_ERROR_NOT_IMPLEMENTED;
+}
 
 }  // namespace net
 }  // namespace mozilla